udpate

+++ c6/numtypes.c
@@ -4,8 +4,8 @@
 
 int main(void) {
   char a;
-  int b;
-  short c;
+  short b;
+  int c;
   long d;
 
   float e;
